% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bridge_sim_loop_cpp}
\alias{bridge_sim_loop_cpp}
\title{Run the Gillespie simulation loop in C++}
\usage{
bridge_sim_loop_cpp(sim_state_r, lambda, rate)
}
\arguments{
\item{sim_state_r}{Named list returned by \code{initialize_simulation()}}

\item{lambda}{Poisson rate for genomic events per daughter}

\item{rate}{Mean of exponential event-length distribution for amp/del}
}
\value{
Named list with same structure as \code{sim_state_r}
}
\description{
Takes an already-initialised \code{sim_state} from R's
\code{initialize_simulation()} and runs the main loop, returning
the final state in the same list format expected by
\code{prepare_results()}.
}
