% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells2countvectors.R
\name{cells2countvectors}
\alias{cells2countvectors}
\title{Convert Cells to Count Vectors}
\usage{
cells2countvectors(cells, L)
}
\arguments{
\item{cells}{A list where each element represents a cell, typically a sequence or other structured data.}

\item{L}{An integer specifying the length or dimension used to determine the breakpoints.}
}
\value{
A matrix where each row represents a count vector for a corresponding cell.
}
\description{
This function converts a list of cells into count vectors based on the breakpoints in the profile.
}
