% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells2mat.R
\name{cells2mat}
\alias{cells2mat}
\title{Convert Cells to a Matrix}
\usage{
cells2mat(cells, L, order)
}
\arguments{
\item{cells}{A list where each element represents a cell, typically a sequence or other structured data.}

\item{L}{An integer specifying the number of columns (length or dimension) of the resulting matrix.}

\item{order}{A logical value indicating whether to reorder rows based on the proportion of entries with a value of 1.}
}
\value{
A matrix where rows represent cells and columns represent counts at each position. If \code{order} is \code{TRUE}, rows are sorted by the proportion of positions with a value of 1.
}
\description{
This function converts a list of cells into a matrix representation where each row corresponds to a cell and columns represent counts at each position.
}
