% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_distance.R
\name{cells_distance}
\alias{cells_distance}
\title{Compute Distance Between Two Sets of Cells}
\usage{
cells_distance(x1, x2, order)
}
\arguments{
\item{x1}{A list representing the first set of cells. It must contain:
\itemize{
\item cells : A list of integer vectors, where each vector represents a cell sequence.
\item input_parameters : A list containing at least the field initial_sequence_length, specifying the expected length of each cell sequence.
}}

\item{x2}{A list representing the second set of cells, structured identically to \code{x1}.}

\item{order}{Logical value indicating whether to reorder cells based on the proportion of genomic regions with exactly one copy.}
}
\value{
A numeric value representing the normalized Euclidean distance between the two sets of cells.
}
\description{
This function calculates the Euclidean distance between two sets of cells, optionally ordering them based on the proportion of the genome in a single-copy state.
}
