% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{continue_simulation}
\alias{continue_simulation}
\title{Check if simulation should continue}
\usage{
continue_simulation(state, max_time, max_cells)
}
\arguments{
\item{state}{The simulation state}

\item{max_time}{Maximum simulation time}

\item{max_cells}{Maximum number of cells allowed}
}
\value{
Logical indicating whether simulation should continue
}
\description{
Check if simulation should continue
}
