% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_tree_annotation}
\alias{create_tree_annotation}
\title{Create tree annotation for heatmap}
\usage{
create_tree_annotation(tree_ggplot, tree_width, n_cells)
}
\arguments{
\item{tree_ggplot}{ggplot tree object}

\item{tree_width}{Width of tree annotation}

\item{n_cells}{Number of cells in heatmap}
}
\value{
HeatmapAnnotation object
}
\description{
Create tree annotation for heatmap
}
\keyword{internal}
