% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fuse_sequence_cpp}
\alias{fuse_sequence_cpp}
\title{Fuse a sequence with its reverse (BFB palindrome creation)}
\usage{
fuse_sequence_cpp(seq_list)
}
\arguments{
\item{seq_list}{R list of intervals}
}
\value{
R list of fused intervals
}
\description{
Fuse a sequence with its reverse (BFB palindrome creation)
}
