% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_breakpoints.R
\name{get_breakpoints}
\alias{get_breakpoints}
\title{Obtain Breakpoints in the Profile}
\usage{
get_breakpoints(cells, L)
}
\arguments{
\item{cells}{A list or other structure representing the cell data to be converted into a matrix.}

\item{L}{An integer specifying the length or dimension required to construct the matrix.}
}
\value{
A numeric vector containing the indices of the breakpoints where the profile changes.
}
\description{
This function identifies the breakpoints in a profile by comparing consecutive columns in a matrix representation of the input cells.
}
