% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gain_loss_profile.R
\name{get_gain_loss_profile}
\alias{get_gain_loss_profile}
\title{Compute Gain and Loss Profiles}
\usage{
get_gain_loss_profile(cells, L)
}
\arguments{
\item{cells}{A list where each element represents a cell, typically a sequence or other structured data.}

\item{L}{An integer specifying the number of columns (length or dimension) of the resulting matrix.}
}
\value{
A list with two elements:
\item{Loss}{A numeric vector representing the loss profile, calculated as the negative mean of positions with values less than 1.}
\item{Gain}{A numeric vector representing the gain profile, calculated as the mean of positions with values greater than 1.}
}
\description{
This function calculates the gain and loss profiles from a list of cells by analyzing the frequency of counts at each position.
}
