% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seq_length_stats.R
\name{get_seq_length_stats}
\alias{get_seq_length_stats}
\title{Compute Sequence Length Statistics}
\usage{
get_seq_length_stats(sim)
}
\arguments{
\item{sim}{A list representing the simulation output, expected to contain a component \code{final_cells} with sequence data.}
}
\value{
A list with the following components:
\item{mean}{The mean of the sequence lengths.}
\item{median}{The median of the sequence lengths.}
\item{sd}{The standard deviation of the sequence lengths.}
}
\description{
This function calculates statistical summaries (mean, median, and standard deviation) of sequence lengths from a simulation object.
}
