% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bridge_sim.R
\name{initialize_without_bfb}
\alias{initialize_without_bfb}
\title{Initialize simulation without BFB for initial cells (diploid version)}
\usage{
initialize_without_bfb(state, initial_sequences)
}
\arguments{
\item{state}{The simulation state}

\item{initial_sequences}{List of initial chromosome sequences}
}
\value{
Updated simulation state
}
\description{
Initialize simulation without BFB for initial cells (diploid version)
}
\keyword{internal}
