% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hotspot_gained.R
\name{is_hotspot_gained}
\alias{is_hotspot_gained}
\title{Check if a Hotspot is Gained}
\usage{
is_hotspot_gained(cell, hotspot)
}
\arguments{
\item{cell}{A sequence or structure representing the cell to be analyzed.}

\item{hotspot}{A vector of positions representing the hotspot to check for gain.}
}
\value{
A logical value: \code{TRUE} if all hotspot positions are present with counts greater than one, otherwise \code{FALSE}.
}
\description{
This function checks whether all positions in a specified hotspot are present with a count greater than one in a given cell.
}
