% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_seqs2CNruns.R
\name{multi_seqs2CNruns}
\alias{multi_seqs2CNruns}
\title{Process multiple sequences (cells) to get copy number runs}
\usage{
multi_seqs2CNruns(sequences, L = NULL)
}
\arguments{
\item{sequences}{A list of sequences, each represented as a list of segments}

\item{L}{Maximum position to consider}
}
\value{
A tibble with copy number runs for all sequences
}
\description{
Process multiple sequences (cells) to get copy number runs
}
