% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{optimize_tree_ordering}
\alias{optimize_tree_ordering}
\title{Optimize tree tip ordering based on distance matrix}
\usage{
optimize_tree_ordering(tree, distance_matrix)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{distance_matrix}{Distance matrix for optimization}
}
\value{
Reordered phylogenetic tree
}
\description{
Optimize tree tip ordering based on distance matrix
}
\keyword{internal}
