% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gain_loss_fraction_profile.R
\name{plot_gain_loss_fraction_profile}
\alias{plot_gain_loss_fraction_profile}
\title{Plot Gain and Loss Fraction Profile}
\usage{
plot_gain_loss_fraction_profile(cells, L)
}
\arguments{
\item{cells}{A list of cell sequences to be analyzed.}

\item{L}{An integer specifying the length of the chromosome or genome being analyzed.}
}
\value{
A ggplot object displaying the gain and loss fractions across chromosome bins.
}
\description{
This function visualizes the gain and loss fraction profile across chromosome bins for a given set of cells.
}
