% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_CN_profiles.R
\name{plot_heatmap_CN_profiles}
\alias{plot_heatmap_CN_profiles}
\title{Plot Heatmap of Copy Number Profiles}
\usage{
plot_heatmap_CN_profiles(cells, L, order = TRUE)
}
\arguments{
\item{cells}{A list of cell sequences to be analyzed.}

\item{L}{An integer specifying the length of the chromosome or genome being analyzed.}

\item{order}{A logical value indicating whether to order cells by their diploid percentage (default is \code{TRUE}).}
}
\value{
A ggplot object displaying the copy number profiles as a heatmap.
}
\description{
This function generates a heatmap visualization of copy number (CN) profiles across chromosome bins for a set of cells.
}
