% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hotspot_copies.R
\name{plot_hotspot_copies}
\alias{plot_hotspot_copies}
\title{Plot the distribution of hotspot copies in a dataset}
\usage{
plot_hotspot_copies(x)
}
\arguments{
\item{x}{A list containing simulation data. It must have elements \code{input_parameters}
and \code{cell_history}, where \code{cell_history} includes \code{hotspot_copies} and \code{is_alive}.}
}
\value{
A ggplot object showing the histogram of hotspot copy numbers, or NULL if the dataset
does not include a hotspot.
}
\description{
This function generates a histogram of hotspot copy numbers from a given dataset.
It filters the dataset to include only living cells and plots the distribution
of hotspot copy numbers.
}
