% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_length_distribution.R
\name{plot_sequence_length_distribution}
\alias{plot_sequence_length_distribution}
\title{Plot Sequence Length Distribution}
\usage{
plot_sequence_length_distribution(
  L0,
  n,
  support,
  Nrep = 1000,
  bins = 100,
  alpha = NULL,
  beta = NULL
)
}
\arguments{
\item{L0}{An integer specifying the initial sequence length.}

\item{n}{An integer representing the number of breakage-fusion-bridge cycles to simulate.}

\item{support}{A character string indicating the type of distribution to use ("uniform" or "beta").}

\item{Nrep}{An integer specifying the number of repetitions for the simulation (default: 1000).}

\item{bins}{An integer defining the number of bins for the histogram (default: 100).}

\item{alpha}{Optional parameter for the alpha value of the beta distribution (if \code{support == "beta"}).}

\item{beta}{Optional parameter for the beta value of the beta distribution (if \code{support == "beta"}).}
}
\value{
A ggplot object displaying the sequence length distribution as a histogram.
}
\description{
This function visualizes the distribution of sequence lengths after a specified number of breakage-fusion-bridge (BFB) cycles.
}
