% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_heatmap_matrix}
\alias{prepare_heatmap_matrix}
\title{Prepare matrix data for heatmap plotting}
\usage{
prepare_heatmap_matrix(
  data,
  chromosomes_to_plot,
  feature_name,
  ordered_cell_ids
)
}
\arguments{
\item{data}{Input data frame}

\item{chromosomes_to_plot}{Chromosomes to include}

\item{feature_name}{Name of feature column to extract}

\item{ordered_cell_ids}{Cell IDs in desired order}
}
\value{
List containing matrix and column information
}
\description{
Prepare matrix data for heatmap plotting
}
\keyword{internal}
