% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_of_Ln.R
\name{probability_of_Ln}
\alias{probability_of_Ln}
\title{Calculate Probability of Final Sequence Length After BFB Cycles}
\usage{
probability_of_Ln(Ln, L0, n)
}
\arguments{
\item{Ln}{Final sequence length after n BFB cycles}

\item{L0}{Initial sequence length before BFB cycles}

\item{n}{Number of BFB cycles}
}
\value{
Probability value between 0 and 1
}
\description{
Calculate Probability of Final Sequence Length After BFB Cycles
}
\details{
Calculates theoretical probability of observing length Ln after n BFB cycles
starting from initial length L0. Returns 0 if Ln exceeds maximum possible length.
}
