% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{process_death_event}
\alias{process_death_event}
\title{Process a death event}
\usage{
process_death_event(state, current_cell_id)
}
\arguments{
\item{state}{The simulation state}

\item{current_cell_id}{ID of the cell undergoing death}
}
\value{
Updated simulation state
}
\description{
Process a death event
}
