% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{process_tree}
\alias{process_tree}
\title{Process tree for heatmap visualization}
\usage{
process_tree(
  tree,
  branch_length = NULL,
  ladderize = FALSE,
  reorder_tree = FALSE,
  distance_matrix = NULL
)
}
\arguments{
\item{tree}{Phylogenetic tree object}

\item{branch_length}{Uniform branch length (optional)}

\item{ladderize}{Whether to ladderize the tree}

\item{reorder_tree}{Whether to optimize tree ordering}

\item{distance_matrix}{Distance matrix for reordering (reorder_tree = TRUE)}
}
\value{
List containing processed tree and ggplot object
}
\description{
Process tree for heatmap visualization
}
\keyword{internal}
