% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2CNruns.R
\name{seq2CNruns}
\alias{seq2CNruns}
\title{Create a tibble of copy number runs from a sequence represented as seq}
\usage{
seq2CNruns(seq, L = NULL)
}
\arguments{
\item{seq}{A list of seq, each with start, end, and direction}

\item{L}{Maximum value to consider (defaults to max value in seq)}
}
\value{
A tibble with columns segment_start, segment_end, and copy_number
}
\description{
Create a tibble of copy number runs from a sequence represented as seq
}
