% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_n_bfb.R
\name{sim_n_bfb}
\alias{sim_n_bfb}
\title{Simulate Breakage-Fusion-Bridge (BFB) Cycles}
\usage{
sim_n_bfb(L, n, support = "uniform", alpha = NULL, beta = NULL)
}
\arguments{
\item{L}{Initial sequence length (number of genomic segments)}

\item{n}{Number of BFB cycles to simulate}

\item{support}{Distribution type for breakpoint selection ("uniform" or "beta")}

\item{alpha}{Shape parameter for beta distribution (only used if support="beta")}

\item{beta}{Shape parameter for beta distribution (only used if support="beta")}
}
\value{
Vector containing the final sequence after n BFB cycles
}
\description{
Simulate Breakage-Fusion-Bridge (BFB) Cycles
}
\details{
For each cycle:
\enumerate{
\item Random breakpoint is selected based on specified distribution
\item Segment left of breakpoint is duplicated and inverted
\item Inverted segment is appended
}
}
