% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{validate_heatmap_inputs}
\alias{validate_heatmap_inputs}
\title{Validate input parameters for heatmap plotting}
\usage{
validate_heatmap_inputs(data, tree, chromosomes_to_plot, to_plot)
}
\arguments{
\item{data}{Input data frame}

\item{tree}{Phylogenetic tree object (optional)}

\item{chromosomes_to_plot}{Vector of chromosomes to include}

\item{to_plot}{Vector of features to plot}
}
\description{
Validate input parameters for heatmap plotting
}
\keyword{internal}
