% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges-package.R
\docType{package}
\name{bridges-package}
\alias{bridges}
\alias{bridges-package}
\title{bridges: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Giovanni Santacatterina \email{first.last@example.com}

}
\keyword{internal}
