% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cells_distance.R
\name{cells_distance}
\alias{cells_distance}
\title{Compute Distance Between Two Sets of Cells}
\usage{
cells_distance(cells1, cells2, L)
}
\arguments{
\item{cells1}{A list of cell sequences representing the first set of cells.}

\item{cells2}{A list of cell sequences representing the second set of cells.}

\item{L}{An integer specifying the length of the vector representation for each cell.}
}
\value{
A list containing:
\item{distance}{The Euclidean distance between the two sets of cells.}
\item{normalized_distance}{The normalized Euclidean distance, scaled by the dimensions of the matrices.}
}
\description{
This function calculates the distance between two sets of cells using a specified distance method.
}
\details{
The function converts each set of cells into a matrix representation, where each row corresponds to a cell, and columns represent bins.
Rows are reordered based on the proportion of features with exactly one copy in descending order. The Euclidean distance between the
two matrices is computed and normalized by the product of the matrix dimensions.
}
