% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{continue_simulation}
\alias{continue_simulation}
\title{Check if simulation should continue}
\usage{
continue_simulation(state)
}
\arguments{
\item{state}{The simulation state}
}
\value{
Logical indicating whether simulation should continue
}
\description{
Check if simulation should continue
}
