% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bridge_sim.R
\name{create_initial_chromosome_sequences}
\alias{create_initial_chromosome_sequences}
\title{Create initial chromosome sequences for all alleles}
\usage{
create_initial_chromosome_sequences(input_parameters)
}
\arguments{
\item{input_parameters}{List of input parameters}
}
\value{
Named list of initial sequences for each chromosome allele
}
\description{
Create initial chromosome sequences for all alleles
}
