% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{get_next_event}
\alias{get_next_event}
\title{Determine the next event to occur}
\usage{
get_next_event(state)
}
\arguments{
\item{state}{The simulation state}
}
\value{
List with time, cell ID, and event type of next event
}
\description{
Determine the next event to occur
}
