% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seq_length.R
\name{get_seq_length}
\alias{get_seq_length}
\title{Compute the Length of a Sequence}
\usage{
get_seq_length(sequence)
}
\arguments{
\item{sequence}{A list of intervals, where each interval is expected to have \code{start}, \code{end}, and \code{direction} components.}
}
\value{
An integer representing the total length of the sequence.
}
\description{
This function calculates the length of a sequence by summing the lengths of its intervals.
}
