% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hotspot_copies_cpp}
\alias{hotspot_copies_cpp}
\title{Count copies of a hotspot bin in an interval-encoded sequence}
\usage{
hotspot_copies_cpp(seq_list, bin)
}
\arguments{
\item{seq_list}{R list of intervals}

\item{bin}{Integer bin position to query}
}
\value{
Integer copy count
}
\description{
Count copies of a hotspot bin in an interval-encoded sequence
}
