% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{initialize_simulation}
\alias{initialize_simulation}
\title{Initialize simulation state}
\usage{
initialize_simulation(
  initial_cells,
  initial_sequence_length,
  birth_rate,
  death_rate,
  positive_selection_rate,
  negative_selection_rate,
  first_round_of_bfb,
  breakpoint_support,
  hotspot,
  alpha,
  beta
)
}
\arguments{
\item{initial_cells}{Number of cells to start with}

\item{initial_sequence_length}{Length of initial genomic sequence}

\item{birth_rate}{Base birth rate}

\item{death_rate}{Base death rate}

\item{positive_selection_rate}{Selection advantage for hotspot cells}

\item{negative_selection_rate}{Selection disadvantage for non hotspot cells}

\item{first_round_of_bfb}{Whether to apply BFB to initial cells}

\item{breakpoint_support}{Distribution for breakpoint selection}

\item{hotspot}{Positions considered hotspots}

\item{alpha}{Beta distribution parameter}

\item{beta}{Beta distribution parameter}
}
\value{
A list containing the initialized simulation state
}
\description{
Initialize simulation state
}
