% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{initialize_simulation}
\alias{initialize_simulation}
\title{Initialize simulation state}
\usage{
initialize_simulation(input_parameters)
}
\arguments{
\item{input_parameters}{List of parameters given as input to gillespie_sim function.}
}
\value{
A list containing the initialized simulation state
}
\description{
Initialize simulation state
}
