% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bridge_sim.R
\name{initialize_simulation}
\alias{initialize_simulation}
\title{Helper functions modified for diploid chromosome modeling
Initialize simulation state for diploid chromosomes}
\usage{
initialize_simulation(input_parameters)
}
\arguments{
\item{input_parameters}{List of parameters given as input to bridge_sim function.}
}
\value{
A list containing the initialized simulation state
}
\description{
Helper functions modified for diploid chromosome modeling
Initialize simulation state for diploid chromosomes
}
