% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{initialize_with_bfb}
\alias{initialize_with_bfb}
\title{Initialize simulation with BFB events for initial cells}
\usage{
initialize_with_bfb(state, initial_sequence)
}
\arguments{
\item{state}{The simulation state}

\item{initial_sequence}{Initial genomic sequence}
}
\value{
Updated simulation state
}
\description{
Initialize simulation with BFB events for initial cells
}
