% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gillespie.R
\name{initialize_without_bfb}
\alias{initialize_without_bfb}
\title{Initialize simulation without BFB for initial cells}
\usage{
initialize_without_bfb(state)
}
\arguments{
\item{state}{The simulation state}
}
\value{
Updated simulation state
}
\description{
Initialize simulation without BFB for initial cells
}
