% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.R
\name{plot_abc_results}
\alias{plot_abc_results}
\title{Plot ABC posterior distributions}
\usage{
plot_abc_results(abc_results)
}
\arguments{
\item{abc_results}{Output of \code{abc_inference()}.}
}
\description{
Plot ABC posterior distributions
}
