% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bridge_sim.R
\name{prepare_results}
\alias{prepare_results}
\title{Prepare final results from simulation}
\usage{
prepare_results(state, f_subsample)
}
\arguments{
\item{state}{The simulation state}

\item{f_subsample}{Fraction of alive cells
to subsample}
}
\value{
A list containing simulation results
}
\description{
Prepare final results from simulation
}
