% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bridge_sim.R
\name{process_birth_event}
\alias{process_birth_event}
\title{Process a birth event (diploid version)}
\usage{
process_birth_event(state, current_cell_id, lambda, rate)
}
\arguments{
\item{state}{The simulation state}

\item{current_cell_id}{ID of the cell undergoing birth}
}
\value{
Updated simulation state
Process a birth event (diploid version)

Updated simulation state
}
\description{
Process a birth event (diploid version)
}
