% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_tree.R
\name{reconstruct_tree}
\alias{reconstruct_tree}
\title{Reconstruct Phylogenetic Tree from Simulation Data}
\usage{
reconstruct_tree(
  sim,
  L,
  alpha = 0.1,
  cn_weight = 0.001,
  eps = 0.01,
  remove_cells = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sim}{A simulation object containing final cells and cell lifetimes.}

\item{L}{Lenght of "normal" sequence}

\item{alpha}{A numeric value controlling the weight of certain genetic features in the distance calculation (default: 0.1).}

\item{cn_weight}{A numeric weight applied to copy number variations in distance calculations (default: 1e-3).}

\item{eps}{A threshold for determining twin cells (default: 0.01).}

\item{remove_cells}{Logical, whether to remove redundant/twin cells (default: TRUE).}

\item{verbose}{Logical, whether to print iteration details (default: FALSE).}
}
\value{
A list containing:
\itemize{
\item \code{cell_df}: A dataframe with inferred cell lineages and events.
\item \code{pd_list}: A list mapping parents to their daughters in the phylogenetic tree.
}
}
\description{
This function reconstructs a phylogenetic tree from a simulation dataset by iteratively clustering cells
based on their genetic similarity. The process continues until a root cell containing all original cells is identified.
}
