% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reverse_sequence_cpp}
\alias{reverse_sequence_cpp}
\title{Reverse an interval-encoded sequence}
\usage{
reverse_sequence_cpp(seq_list)
}
\arguments{
\item{seq_list}{R list of intervals}
}
\value{
R list of reversed intervals
}
\description{
Reverse an interval-encoded sequence
}
