% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq2vec_cpp}
\alias{seq2vec_cpp}
\title{Expand interval-encoded sequence to an integer vector}
\usage{
seq2vec_cpp(seq_list)
}
\arguments{
\item{seq_list}{R list of intervals}
}
\value{
IntegerVector of genomic bin values
}
\description{
Expand interval-encoded sequence to an integer vector
}
