% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_length_cpp}
\alias{seq_length_cpp}
\title{Compute length of an interval-encoded sequence}
\usage{
seq_length_cpp(seq_list)
}
\arguments{
\item{seq_list}{R list of intervals (each a named list with start, end, direction)}
}
\value{
Integer total length
}
\description{
Compute length of an interval-encoded sequence
}
\keyword{internal}
