% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_amp_del_cpp}
\alias{sim_amp_del_cpp}
\title{Simulate amplification or deletion on an interval-encoded sequence}
\usage{
sim_amp_del_cpp(seq_list, operation, rate)
}
\arguments{
\item{seq_list}{R list of intervals}

\item{operation}{\code{"dup"} for duplication or \code{"del"} for deletion}

\item{rate}{Mean of the exponential event-length distribution}
}
\value{
R list of intervals
}
\description{
Simulate amplification or deletion on an interval-encoded sequence
}
