% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bridge_sim.R
\name{sim_bfb_left_and_right_sequences}
\alias{sim_bfb_left_and_right_sequences}
\title{Simulate Breakage-Fusion-Bridge (BFB) Cycle for both daughters}
\usage{
sim_bfb_left_and_right_sequences(
  sequence,
  support = "uniform",
  alpha = NULL,
  beta = NULL
)
}
\arguments{
\item{sequence}{Input sequence}

\item{support}{Distribution type for breakpoint selection ("uniform" or "beta")}

\item{alpha}{Shape parameter for beta distribution (only used if support="beta")}

\item{beta}{Shape parameter for beta distribution (only used if support="beta")}
}
\value{
List containing left and right sequences
}
\description{
Simulate Breakage-Fusion-Bridge (BFB) Cycle for both daughters
}
\details{
Simulate left and right children from a BFB cycle using specified
breakpoint selection distribution
}
