% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec2seq_cpp}
\alias{vec2seq_cpp}
\title{Compress an integer vector to an interval-encoded sequence}
\usage{
vec2seq_cpp(vec)
}
\arguments{
\item{vec}{IntegerVector of genomic bin values}
}
\value{
R list of intervals
}
\description{
Compress an integer vector to an interval-encoded sequence
}
\keyword{internal}
